# Conflicts:
#	backend/routes/api.php

+++ backend/app/Http/Controllers/SkillController.php
@@ -6,7 +6,7 @@ use App\Comments;
 use App\PortfolioWorks;
 use Illuminate\Http\Request;
 use App\Skill;
-use App\SkillApproval;
+use App\SkillUser;
 use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Validator;
 use App\User;
@@ -14,15 +14,8 @@ use Illuminate\Support\Facades\DB;
 
 class SkillController extends Controller
 {
-    public function index()
-    {
-        return DB::table("users")
-            ->join('user_skill','user_skill.user_id','users.id')
-            ->join('skill','skill.id','user_skill.skill_id')
-            ->get();
 
-    }
-    public function create(Request $request, SkillApproval $skillas)
+    public function create(Request $request, SkillUser $skillas)
     {
         try {
             $user = auth()->userOrFail();
@@ -31,49 +24,17 @@ class SkillController extends Controller
         }
         $skillArr = $request->all();
         $loopLenght = count($skillArr);
-        SkillApproval::where('user_id',auth()->user()->id)->delete();
+        SkillUser::where('user_id',auth()->user()->id)->delete();
         for ($i = 0; $i < $loopLenght; $i++){
-                $skill = new SkillApproval;
+                $skill = new SkillUser;
                 $skill->user_id = auth()->user()->id;
                 $skill->skill_id = $skillArr[$i];
                 $skill->approved = 0;
-                $skill->comment = '';
                 $skill->save();
-
         }
 
 
         return response()->json(["message" => "Skill pridėtas"]);
-    }
-
-    public function update(Request $request,SkillApproval $skill)
-    {
-        try {
-            $user = auth()->userOrFail();
-        } catch(\Tymon\JWTAuth\Exceptions\UserNotDefinedException $e) {
-            return response()->json(['error' => 'Prašome prisijungti'], 401);
-        }
-        $skill->approved = $request->approved;
-        $skill->comment = $request->comment;
-        $skill->save();
-
-        return response()->json(["message" => "Skill atnaujintas"]);
-    }
-
-    public function delete(SkillApproval $id)
-    {
-        try {
-            $user = auth()->userOrFail();
-        } catch(\Tymon\JWTAuth\Exceptions\UserNotDefinedException $e) {
-            return response()->json(['error' => 'Prašome prisijungti'], 401);
-        }
-        if (Gate::allows('authorization', $id)) {
-            $id->delete();
-        } else if (Gate::denies('authorization', $id)) {
-            return response()->json(["error" => "Jūs neturite teisės"], 403);
-        }
-        return response()->json(["message" => "Skill istrintas"]);
-
     }
     
     public function skillsList() {
@@ -88,6 +49,22 @@ class SkillController extends Controller
         }
         $skill->delete();
         return response()->json(["message"=>"Ištrinta"],200);
+    }
+    public function aboutSkillUser($id)
+        {
+            return response()->json(SkillUser::select('*')->where('user_id',$id)->get(),200);
+        }
+
+    public function SkillApproval(Request $request, $skill_id, $user_id)
+    {
+        $skill = Skill::where('id', $skill_id)->first();
+        $user = User::where('id', $user_id)->first();
+        $skill->users()->attach($user);
+        $validatedData = $request->validate([
+            'approved' => 'required'
+        ]);
+        SkillUser::where('user_id', $user_id)->where('skill_id', $skill_id)->update(['approved' => $request->input('approved')]);
+        return response()->json(["message" => "Skilas sekmingai patvirtintas",200]);
     }
     
 }
